used goslack instead of smoothoperator due to compatability issues
