--- internal/controller/slack.go
@@ -1,27 +1,0 @@
-package controller
-
-import (
-	"context"
-	smoothoperatorslack "github.com/pdok/smooth-operator/pkg/integrations/slack"
-	"sigs.k8s.io/controller-runtime/pkg/log"
-)
-
-type SlackSender interface {
-	Send(message string, ctx context.Context)
-}
-
-type Slack struct {
-	SlackUrl string
-}
-
-func (s *Slack) Send(message string, ctx context.Context) {
-	const messagePrefix = "ogcapi-operator failed: "
-	lgr := log.FromContext(ctx)
-
-	slackRequest := smoothoperatorslack.GetSimpleSlackErrorMessage(messagePrefix + message)
-
-	err := smoothoperatorslack.SendSlackRequest(slackRequest, s.SlackUrl)
-	if err != nil {
-		lgr.Error(err, "unable to send Slack Error message", "message", message)
-	}
-}
